+++ example/react-example/extend.d.ts
@@ -1,0 +1,5 @@
+declare module 'config' {
+  const content: string;
+  export default content;
+  export const a: number;
+}