+++ example/react-example/src/pages/index/index.tsx
@@ -1,6 +1,9 @@
+import { a } from 'config';
 import * as React from 'react';
 import { hot } from 'react-hot-loader';
 import './index.less';
+declare var c:string
+console.log(c)
 class A extends React.Component {
   public render() {
     return <div>ccvvvc</div>;