+++ packages/plugin-webpack/src/imp/build.ts
@@ -26,6 +26,7 @@ export default class BuildCommand implements Interfaces.Icli<{}> {
   }
   public async handler() {
     const config = chainConfig('production');
+    
     const compiler = webpack(config.toConfig());
     compiler.run((err, webpackStats) => {
       if (err) {