+++ packages/plugin-webpack/src/imp/dev.ts
@@ -1,9 +1,13 @@
 import { Container, Interfaces, utils } from '@ddot/plugin-utils';
 import { ip } from 'address';
+import * as Fastify from 'fastify';
 import { fatal } from 'signale';
+import { PassThrough } from 'stream';
 import * as webpack from 'webpack';
 import * as devMiddleware from 'webpack-dev-middleware';
+import { createGzip } from 'zlib';
 import chainConfig from '../config';
+import { getCfgSetting } from '../utils';
 
 interface IArgv {
   port: number;
@@ -29,7 +33,7 @@ export default class DevCommand implements Interfaces.Icli<IArgv> {
       return;
     }
     const config = chainConfig('development');
-    const fastify = require('fastify')();
+    const fastify = Fastify();
     const hasFriendError = config.plugins.has('friendly-errors');
     if (hasFriendError) {
       config.plugin('friendly-errors').tap(args => [
@@ -47,19 +51,54 @@ export default class DevCommand implements Interfaces.Icli<IArgv> {
         },
       ]);
     }
-    Object.keys(config.entryPoints.entries()).forEach(name => {
-      config.entry(name).add('webpack-hot-middleware/client');
-    });
+    const cfgsetting = getCfgSetting();
+    if (cfgsetting.hot) {
+      Object.keys(config.entryPoints.entries()).forEach(name => {
+        config.entry(name).add('webpack-hot-middleware/client');
+      });
+    }
     const compiler = webpack(config.toConfig());
     const instance = devMiddleware(compiler, {
       logLevel: hasFriendError ? 'silent' : 'info',
     });
     fastify.use(instance);
-    fastify.use(
-      require('webpack-hot-middleware')(compiler, {
-        log: false,
-      })
-    );
+    if (cfgsetting.hot) {
+      fastify.use(
+        require('webpack-hot-middleware')(compiler, {
+          log: false,
+        })
+      );
+    }
+    fastify.setNotFoundHandler((request, reply) => {
+      try {
+        const filename = compiler.options.output.path + '/index.html';
+        if (
+          /get/i.test(request.req.method) &&
+          ['text/html', '*/*'].filter(
+            type => request.headers.accept.indexOf(type) > -1
+          ).length &&
+          instance.fileSystem.statSync(filename).isFile()
+        ) {
+          reply
+            .type('text/html')
+            .send(instance.fileSystem.readFileSync(filename));
+        } else {
+          reply.code(404).send(new Error('Not Found'));
+        }
+      } catch (e) {
+        reply.code(404).send(new Error('Not Found'));
+      }
+    });
+    cfgsetting.fastify(fastify);
+    Object.keys(cfgsetting.proxy).forEach(prefix => {
+      const { target, ...rest } = cfgsetting.proxy[prefix];
+      fastify.register(require('fastify-http-proxy'), {
+        upstream: target,
+        prefix,
+        rewritePrefix: prefix,
+        ...rest,
+      });
+    });
     fastify.listen(port, '0.0.0.0', err => {
       if (err) {
         fatal(err);